+++ packages/docker-image/eslint.config.mjs
@@ -1,3 +1,0 @@
-import { nodeConfig } from "@envpilot/eslint-config/node";
-
-export default nodeConfig;